+++ nqct/models/__init__.py
@@ -2,15 +2,21 @@
 
 from nqct.models.backend import Backend, BackendQueueStatus
 from nqct.models.execution import (
+    AcquisitionType,
+    AveragingMode,
     ExecutionConfig,
     HardwareExecutionConfig,
     QubitMappingEntry,
     SimulatorExecutionConfig,
+    normalize_acquisition_type,
+    normalize_averaging,
 )
 from nqct.models.function import Function
 from nqct.models.job import Job
 
 __all__ = [
+    "AcquisitionType",
+    "AveragingMode",
     "Backend",
     "BackendQueueStatus",
     "ExecutionConfig",
@@ -19,4 +25,6 @@ __all__ = [
     "Job",
     "QubitMappingEntry",
     "SimulatorExecutionConfig",
+    "normalize_acquisition_type",
+    "normalize_averaging",
 ]